+++ config/bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 use Cake\Core\Configure;
 use Cake\Event\EventManager;
-use Search\Event\WidgetsListener;
+use Search\Event\Model\WidgetsListener;
 
 // dashboards columns
 Configure::write('Search.dashboard.columns', ['Left Side', 'Right Side']);
